Defined connectivity service exceptions in separate file
