+++ src/drunc/connectivity_service/exceptions.py
@@ -1,0 +1,14 @@
+from drunc.exceptions import DruncException
+
+
+class ApplicationRegistryNotPresent(DruncException):
+    pass
+
+class ApplicationRegistrationUnsuccessful(DruncException):
+    pass
+
+class ApplicationLookupUnsuccessful(DruncException):
+    pass
+
+class ApplicationUpdateUnsuccessful(DruncException):
+    pass
